feat: navigate by arrow keys instead of wasd in pathoscope detailed view
